fix: first task delete
