+++ app/src/main/res/drawable/round_priority_high_24.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:tint="#000000"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M8,19m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0" />
+
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M8,3c-1.1,0 -2,0.9 -2,2v8c0,1.1 0.9,2 2,2s2,-0.9 2,-2V5c0,-1.1 -0.9,-2 -2,-2z" />
+
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M16,19m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0" />
+
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M16,3c-1.1,0 -2,0.9 -2,2v8c0,1.1 0.9,2 2,2s2,-0.9 2,-2V5c0,-1.1 -0.9,-2 -2,-2z" />
+
+</vector>
